mostly configures

+++ src/config.rs
@@ -72,16 +72,16 @@ pub const USB_PULLUP_ACTVE_LEVEL: bool = false;
 //-----------------------------------------------------------------------------
 
 /// max laser S 100 -> 100%
-pub const MOTION_MAX_S: f64 = 100f64;
+pub const MOTION_MAX_S: f32 = 100f32;
 
 /// working range X
-pub const MOTION_X_RANGE: f64 = 250.0;
+pub const MOTION_X_RANGE: f32 = 250.0;
 
 /// working range Y
-pub const MOTION_Y_RANGE: f64 = 250.0;
+pub const MOTION_Y_RANGE: f32 = 250.0;
 
 /// working range Z (unused)
-pub const MOTION_Z_RANGE: f64 = 1.0;
+pub const MOTION_Z_RANGE: f32 = 1.0;
 
 /// invert axis
 pub const AXIS_INVERSE_X: bool = false;
